FIX Implement TestOnly in PretendLinkChecker, move Page stub and sort class imports
